Move pairing functions to their own package
